file handlers and bug fixes
getItemsInfo: returns a string object to contains all the nessesary details for the assignment so that it can be pushed forward directly to the file

saveMenuToFolder: takes a vector of items. Make sure they're sorted to what we need.
The outputFileHandler simply needs to be initiated with where to write, as well as needs to be closed after

saveMenuToFolder's overloaded method is used to cross check the output from vector of items that you give it

Also, refactored getItems and fixed it so that it dosn't return any empty items. Important fix as it usually ends up breaking the code otherwise, but everything should work fine from now